ignore download if file exits on linux

+++ download-extra-dataset.sh
@@ -1,12 +1,17 @@
 #!/bin/bash
 
-echo -e "Downloading Los Angeles crime data from data.gov"
-curl -s -o ./dataset/csv/crime_data.csv https://data.lacity.org/api/views/2nrs-mtv8/rows.csv\?accessType\=DOWNLOAD
-
-if [ $? -eq 0 ]
+if ! [ -f ./dataset/csv/crime_data.csv ]
 then
-  echo "Download complete"
+	echo -e "Downloading Los Angeles crime data from data.gov"
+	curl -s -o ./dataset/csv/crime_data.csv https://data.lacity.org/api/views/2nrs-mtv8/rows.csv\?accessType\=DOWNLOAD
+
+	if [ $? -eq 0 ]
+	then
+		echo "Download complete"
+	else
+		echo "Download failed"
+		exit 1
+	fi
 else
-  echo "Download failed"
-  exit 1
+	echo -e "crime data already exists"
 fi
